+++ src/candy/combo/Comboable.hpp
@@ -7,12 +7,12 @@
 class Comboable : public Valuable {
 protected:
     int level;
-    double comboValue;
+    unsigned long long comboValue;
 
 public:
     explicit Comboable(int level);
 
-    Comboable(int level, double comboValue);
+    Comboable(int level, unsigned long long int comboValue);
 
     [[nodiscard]] value_pair_t value() const override;
 };
